% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{anchor_asq}
\alias{anchor_asq}
\title{Prices of 50,000 round cut diamonds}
\format{A data frame with 29 rows and 8 variables}
\usage{
data(anchor_asq)
}
\description{
A dataset containing the responses of prices and other attributes of almost 54,000
 diamonds. The variables are as follows:
}
\details{
\itemize{
  \item item_order. index of the item, same to those used in \code{itemmap_asq} data.
  \item item_id. item ID, same to those in \code{response_asq} data.
  \item a. the discrimination (slope) parameter.
  \item cb1 - cb4. the boundaries between each category-pair.
  \item NCAT. the number of categories.
}
}
\keyword{datasets}
