% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{checkFilePath}
\alias{checkFilePath}
\title{Check file path}
\usage{
checkFilePath(abspath, path)
}
\arguments{
\item{abspath}{Absolute path of an input directory.}

\item{path}{Absolute or relative path of a file.}
}
\value{
a list containing normalized file path and whether it exists.
}
\description{
Check file path to see if it exists.
}
