% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{compareScores}
\alias{compareScores}
\title{Compare two sets of scores}
\usage{
compareScores(left, right, type = c("corr", "mean", "sd", "rmsd"))
}
\arguments{
\item{left}{Scores on the left side of comparison.}

\item{right}{Scores on the right side of comparison. This is subtracted from 'left'.}

\item{type}{Type of comparisons to include. Accepts 'corr', "mean', 'sd', 'rmsd'. Defaults to all four.}
}
\description{
Compare two sets of scores.
}
