% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_dep}
\alias{dataset_dep}
\alias{response_dep}
\alias{anchor_dep}
\alias{itemmap_dep}
\alias{data_dep}
\title{DEP dataset}
\description{
This dataset is associated with the following objects:
}
\details{
\itemize{
  \item \code{response_dep} Raw response matrix. A data frame with 747 rows and 49 variables.
  \itemize{
    \item \code{prosettaid}. Sample IDs.
    \item \code{EDDEP04 -- EDDEP54}. Response to anchor items.
    \item \code{CESD1 -- CESD20}. Response to development items.
  }
  \item \code{itemmap_dep} Item map describing the items in each instrument. A data frame with 48 rows and 9 variables.
  \itemize{
    \item \code{item_order} Item index. Refers to column \code{item_order} in anchor item attributes.
    \item \code{instrument} Instrument index.
    \item \code{item_id} Item ID. Refers to column \code{prosettaid} in \code{response_asq} data.
    \item \code{item_name} New item names to be used in mapping.
    \item \code{NCAT} The number of response categories.
    \item \code{minScore} The minimum score of the item.
    \item \code{reverse} 1 indicates a reverse-scored item.
    \item \code{scores} A vector representing all possible score values from the item.
  }
  \item \code{anchor_dep} Anchor item attributes. A data frame with 28 rows and 8 variables.
  \itemize{
    \item \code{item_order} Item index.
    \item \code{item_id} Item ID, same to those in the response data above.
    \item \code{a} The discrimination (slope) parameter.
    \item \code{cb1 - cb4} The boundaries between each category-pair.
    \item \code{NCAT} The number of response categories.
  }
}
}
\examples{
\dontrun{
response_dep
itemmap_dep
anchor_dep
data_dep
}

data_dep <- loadData(
  response = response_dep,
  itemmap  = itemmap_dep,
  anchor   = anchor_dep
)
}
\keyword{datasets}
