% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{detectParameterization}
\alias{detectParameterization}
\title{(internal) detect parameterization type}
\usage{
detectParameterization(ipar)
}
\arguments{
\item{ipar}{a \code{\link{data.frame}} containing item parameters.}
}
\value{
\code{\link{detectParameterization}} returns \code{ab} or \code{ad}.
}
\description{
\code{\link{detectParameterization}} is an internal function for detecting the type of parameterization used in a set of item parameters.
}
\examples{
PROsetta:::detectParameterization(data_asq@anchor) # ab

}
\keyword{internal}
