% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{getCompleteData}
\alias{getCompleteData}
\title{Get complete data}
\usage{
getCompleteData(data)
}
\arguments{
\item{data}{A \code{\linkS4class{PROsetta_data}} object.}
}
\description{
Perform casewise deletion of missing values.
}
