% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{getEscore}
\alias{getEscore}
\title{Computed expected scores at thetas using item parameters}
\usage{
getEscore(ipar, model, theta, is_minscore_0)
}
\arguments{
\item{ipar}{Item parameters}

\item{model}{Item model to use}

\item{theta}{Theta values}

\item{is_minscore_0}{If TRUE the score begins from 0 instead of 1.}
}
\description{
Computed expected scores at thetas using item parameters
}
