% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{getRSSS}
\alias{getRSSS}
\title{Compute a Crosswalk Table}
\usage{
getRSSS(ipar, theta_grid, is_minscore_0, prior_mu_sigma)
}
\arguments{
\item{ipar}{an item parameter matrix for graded response items. Accepts both a/b and a/d format parameters. Accepts multidimensional item parameters.}

\item{theta_grid}{the theta grid to use.}

\item{is_minscore_0}{if \code{TRUE}, the scores of each item begins from 0. if \code{FALSE}, the scores of each item begins from 1.}

\item{prior_mu_sigma}{a named list containing prior distribution parameters:
\itemize{
  \item{\code{mu} means}
  \item{\code{sigma} the covariance matrix}
  \item{\code{sd} standard deviations}
  \item{\code{corr} the correlation matrix}
}}
}
\description{
\code{\link{getRSSS}} is a function to generate a raw-score to standard-score crosswalk table.
}
