% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{getScaleSum}
\alias{getScaleSum}
\title{Compute raw sum scores of the specified scale}
\usage{
getScaleSum(data, scale_idx)
}
\arguments{
\item{data}{A \code{\linkS4class{PROsetta_data}} object.}

\item{scale_idx}{The index of the scacle to obtain the raw sum scores.}
}
\description{
Compute raw sum scores of the specified scale
}
