% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{getTheta}
\alias{getTheta}
\title{Obtain EAP estimates}
\usage{
getTheta(
  data,
  ipar,
  scale = "combined",
  model = "grm",
  theta_grid = seq(-4, 4, 0.1),
  prior_dist = "normal",
  prior_mean = 0,
  prior_sd = 1
)
}
\arguments{
\item{data}{A \code{\linkS4class{PROsetta_data}} object.}

\item{ipar}{A data frame containing item parameters.}

\item{scale}{The index of the scale to use. Set to 'combined' to use the combined scale.}

\item{model}{The item model to use. Accepts 'grm' or 'gpcm'.}

\item{theta_grid}{The theta grid to use in calculating EAP estimates.}

\item{prior_dist}{The type of prior distribution. Accepts 'normal' or 'logistic'.}

\item{prior_mean}{Mean of the prior distribution.}

\item{prior_sd}{SD of the prior distribution.}
}
\description{
Obtain EAP estimates
}
