% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshiny.R
\name{guiPROsetta}
\alias{guiPROsetta}
\title{guiPROsetta}
\usage{
guiPROsetta()
}
\description{
\code{\link{guiPROsetta}} is a caller function to launch a Shiny app locally.
}
\examples{
\donttest{
guiPROsetta()
}

}
