% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{itemmap_asq}
\alias{itemmap_asq}
\title{Prices of 50,000 round cut diamonds}
\format{A data frame with 40 rows and 9 variables}
\usage{
data(itemmap_asq)
}
\description{
A dataset containing the responses of prices and other attributes of almost 54,000
 diamonds. The variables are as follows:
}
\details{
\itemize{
  \item item_order. index of the item.
  \item instrument. index of the instrument.
  \item item_id. item ID, same to those in \code{response_asq} data.
  \item item_name. new item names to use in mapping.
  \item NCAT. number of categories.
  \item minScore. minimum score in the item.
  \item reverse. \code{1} indicates a reverse item. accepts (0, 1)
  \item scores. a vector representing all possible values in an item.
  \item NCATc. (placeholder)
}
}
\keyword{datasets}
