% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{LoadData}
\alias{LoadData}
\title{Load data from supplied config}
\usage{
LoadData(Config)
}
\arguments{
\item{Config}{A PROsetta.Config object. See \linkS4class{PROsetta.Config}.}
}
\value{
An S4-class object containing the loaded data.
}
\description{
Loads data from the supplied configuration.
}
