% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{new.config}
\alias{new.config}
\title{Creates a config object for PROsetta}
\usage{
new.config(studyName = "Study", inputDirectory = getwd(),
  outputDirectory = getwd(), itemID = "", personID = "",
  scaleID = "", responseFile, itemmapFile, anchorFile,
  linkingMethod = "FIXEDPAR", guessID = F)
}
\arguments{
\item{studyName}{Character. A user-specified name of the study.}

\item{inputDirectory}{Character. The path of the input directory for reading dataset. Defaults to \code{getwd()}.}

\item{outputDirectory}{Character. The path of the output directory for writing outputs. Defaults to \code{getwd()}. A new folder is created at the path if it does not exist.}

\item{itemID}{Character. The name of the variable representing item IDs in \code{anchorFile} and \code{itemmapFile}.}

\item{personID}{Character. The name of the variable representing person IDs in \code{responseFile}.}

\item{scaleID}{Character. The name of the variable representing scale IDs in \code{itemmapFile}.}

\item{responseFile}{Character. The filename of a dataset containing IDs and the responses of the items.}

\item{itemmapFile}{Character. The filename for an item map.}

\item{anchorFile}{Character. The filename of a dataset containing the parameters of the items.}

\item{linkingMethod}{Character. The linking method to use. Accepts "\code{MM}", "\code{MS}", "\code{HB}", "\code{SL}", "\code{FIXEDPAR}", or "\code{NONE}".}
}
\value{
A \code{\linkS4class{PROsetta.Config}} object.
}
\description{
Creates a config object for PROsetta
}
