% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\docType{methods}
\name{plot,PROsetta_data,ANY-method}
\alias{plot,PROsetta_data,ANY-method}
\title{Plot from objects}
\usage{
\S4method{plot}{PROsetta_data,ANY}(
  x,
  y,
  scale_id = 0,
  filename = NULL,
  title = NULL,
  xlim = NULL,
  color = "blue",
  nbar = 20,
  rug = FALSE,
  filetype = "pdf",
  savefile = FALSE,
  bg = "white",
  width = 6,
  height = 6,
  pointsize = 12
)
}
\arguments{
\item{x}{A \code{\linkS4class{PROsetta_data}} object.}

\item{y}{Unused argument, exists for compatibility with \code{\link{plot}} in the base R package.}

\item{scale_id}{Instrument ID to plot. 0 represents the combined scale.}

\item{filename}{Filename to write if 'savefile' is TRUE.}

\item{title}{The title of the figure.}

\item{xlim}{The range of scores to plot.}

\item{color}{The color to fill the histogram.}

\item{nbar}{The number of histogram bars.}

\item{rug}{If TRUE, display the actual distribution of scores below each bar.}

\item{filetype}{The type of file to write if 'savefile' is TRUE. Accepts 'pdf', 'jpeg', 'png', and 'tiff'.}

\item{savefile}{If TRUE, save the figure as a file.}

\item{bg}{The background color of the plot.}

\item{width}{The width of the plot.}

\item{height}{The height of the plot.}

\item{pointsize}{Point size to pass onto file writing functions.}
}
\description{
Plot from objects in the PROsetta package.
}
\examples{

\donttest{
cfg <- createConfig(
  response_file = "inst/data-raw/dat_decesd_v2.csv",
  anchor_file = "inst/data-raw/anchor_decesd.csv",
  itemmap_file = "inst/data-raw/imap_decesd.csv"
)
plot(cfg)
}

}
