% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\docType{class}
\name{PROsetta.Config-class}
\alias{PROsetta.Config-class}
\title{An S4 class to represent configurations for reading datasets and writing outputs.}
\description{
An S4 class to represent configurations for reading datasets and writing outputs.
}
\section{Slots}{

\describe{
\item{\code{studyName}}{Character. A user-specified name of the study.}

\item{\code{inputDirectory}}{Character. The path of the input directory for reading dataset.}

\item{\code{outputDirectory}}{Character. The path of the output directory for writing outputs. A new folder is created at the path if it does not exist.}

\item{\code{itemID}}{Character. The name of the variable representing item IDs in \code{anchorFile} and \code{itemmapFile}.}

\item{\code{personID}}{Character. The name of the variable representing person IDs in \code{responseFile}.}

\item{\code{scaleID}}{Character. The name of the variable representing scale IDs in \code{itemmapFile}.}

\item{\code{responseFile}}{Character. The filename of a dataset containing IDs and the responses of the items.}

\item{\code{itemmapFile}}{Character. The filename for an item map.}

\item{\code{anchorFile}}{Character. The filename of a dataset containing the parameters of the items.}

\item{\code{linkingMethod}}{Character. The linking method to use. Accepts "\code{MM}", "\code{MS}", "\code{HB}", "\code{SL}", "\code{FIXEDPAR}", or "\code{NONE}".}
}}

\examples{

\dontrun{
cfg <- new.config(
  responseFile = "data-raw/dat_axmasq_v2.csv",
  anchorFile = "data-raw/anchor_axmasq.csv",
  itemmapFile = "data-raw/imap_axmasq.csv",
  linkingMethod = "FIXEDPAR",
  guessID = TRUE
)
}
}
