% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\docType{class}
\name{PROsetta.Data-class}
\alias{PROsetta.Data-class}
\title{An S4 class to represent \code{PROsetta} datasets.}
\description{
An S4 class to represent \code{PROsetta} datasets.
}
\section{Slots}{

\describe{
\item{\code{response}}{A list containing IDs and the responses of the items.}

\item{\code{itemmap}}{A list containing an item map.}

\item{\code{anchor}}{A list containing the parameters of the items.}
}}

