% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshiny.R
\name{PROsetta}
\alias{PROsetta}
\title{PROsetta}
\usage{
PROsetta()
}
\description{
\code{\link{PROsetta}} is a caller function for launching a Shiny app locally.
}
\examples{
if (interactive()) {
  PROsetta()
}

}
