% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{response_asq}
\alias{response_asq}
\title{Prices of 50,000 round cut diamonds}
\format{A data frame with 751 rows and 41 variables}
\usage{
data(response_asq)
}
\description{
A dataset containing the responses of prices and other attributes of almost 54,000
 diamonds. The variables are as follows:
}
\details{
\itemize{
  \item prosettaid. sample ID
  \item EDANX01 -- EDANX55. response to items
  \item MASQ1 -- MASQ11. response to items
}
}
\keyword{datasets}
