% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{RunFrequency}
\alias{RunFrequency}
\title{Obtain a frequency table}
\usage{
RunFrequency(Config, Data = NULL, checkFrequency = T)
}
\arguments{
\item{Config}{A PROsetta.Config object. See \code{\linkS4class{PROsetta.Config}}.}

\item{Data}{A PROsetta.Data object. See \code{\link{LoadData}} for loading a dataset.}

\item{checkFrequency}{Logical. If \code{TRUE}, check the frequency table for missing response categories, and display warning message if any is missing.}
}
\value{
A \code{data.frame} containing the frequency table of the dataset.
}
\description{
Obtains a frequency table from the supplied configuration and the dataset.
}
\examples{
\dontrun{
inputData = new("PROsetta.Data")
inputData@anchor = anchor_asq
inputData@response = response_asq
inputData@itemmap = itemmap_asq
freqTable = RunFrequency(new.Config, inputData)
}
}
