% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{checkFrequency}
\alias{checkFrequency}
\title{Check frequency table for unobserved response categories}
\usage{
checkFrequency(data)
}
\arguments{
\item{data}{A \code{\linkS4class{PROsetta_data}} object. See \code{\link{loadData}} for loading a dataset.}
}
\value{
\code{TRUE} if all categories are present. \code{FALSE} otherwise.
}
\description{
Checks frequency table for unobserved response categories.
}
