% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{CheckFrequency}
\alias{CheckFrequency}
\title{Check frequency table for unobserved response categories}
\usage{
CheckFrequency(Config, Data = NULL)
}
\arguments{
\item{Config}{A PROsetta.Config object. See \code{\linkS4class{PROsetta.Config}}.}

\item{Data}{A PROsetta.Data object. See \code{\link{LoadData}} for loading a dataset.}
}
\value{
Logical. \code{TRUE} if all categories are present. \code{FALSE} otherwise.
}
\description{
Checks frequency table for unobserved response categories.
}
