% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{checkFrequency}
\alias{checkFrequency}
\title{Check frequency table for unobserved response categories}
\usage{
checkFrequency(data)
}
\arguments{
\item{data}{a \code{\linkS4class{PROsetta_data}} object. See \code{\link{loadData}} for loading a dataset.}
}
\value{
If all response categories have a frequency of at least 1, the value is \code{TRUE}.

Otherwise, the value is \code{FALSE}.
}
\description{
\code{\link{checkFrequency}} is a descriptive function to check whether all response categories in a frequency table have a frequency of at least 1.
}
