% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{compareScores}
\alias{compareScores}
\title{Compare two sets of scores}
\usage{
compareScores(left, right, type = c("corr", "mean", "sd", "rmsd", "mad"))
}
\arguments{
\item{left}{scores on the left side of comparison.}

\item{right}{scores on the right side of comparison. This is subtracted from 'left'.}

\item{type}{type of comparisons to include. Accepts '\code{corr}', "\code{mean}', '\code{sd}', '\code{rmsd}'. Defaults to all four.}
}
\value{
\code{\link{compareScores}} returns a \code{\link{data.frame}} containing the comparison results.
}
\description{
\code{\link{compareScores}} is a helper function to compare two sets of scores.
}
