% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getCompleteData}
\alias{getCompleteData}
\title{Get complete data}
\usage{
getCompleteData(data, scale = NULL, verbose = FALSE)
}
\arguments{
\item{data}{a \code{\linkS4class{PROsetta_data}} object.}

\item{scale}{the index of the scale to perform casewise deletion. Leave empty or set to "combined" to perform on all scales.}

\item{verbose}{if \code{TRUE}, print status messages. (default = \code{FALSE})}
}
\value{
\code{\link{getCompleteData}} returns an updated \code{\linkS4class{PROsetta_data}} object.
}
\description{
\code{\link{getCompleteData}} is a helper function for performing casewise deletion of missing values.
}
\examples{
d <- getCompleteData(data_asq, verbose = TRUE)

}
