% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{getEscore}
\alias{getEscore}
\title{Calculate expected scores at theta}
\usage{
getEscore(ipar, model, theta, is_minscore_0)
}
\arguments{
\item{ipar}{item parameters.}

\item{model}{item model to use.}

\item{theta}{theta values.}

\item{is_minscore_0}{if \code{TRUE} the score begins from 0 instead of 1.}
}
\value{
\code{\link{getEscore}} returns a vector of expected scores.
}
\description{
\code{\link{getEscore}} is a helper function for obtaining expected scores at supplied thetas.
}
