% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getItemNames}
\alias{getItemNames}
\title{Get item names}
\usage{
getItemNames(d, scale_id)
}
\arguments{
\item{d}{a \code{\linkS4class{PROsetta_data}} object.}

\item{scale_id}{scale IDs to extract item names.}
}
\value{
\code{\link{getItemNames}} returns a vector containing item names.
}
\description{
\code{\link{getItemNames}} is a helper function for extracting item names for a specified scale from a \code{\linkS4class{PROsetta_data}} object.
}
\examples{
getItemNames(data_asq, 1)
getItemNames(data_asq, 2)
getItemNames(data_asq, c(1, 2))
getItemNames(data_asq, c(2, 1))

}
