% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{getScaleSum}
\alias{getScaleSum}
\title{Calculate raw sum scores of a scale}
\usage{
getScaleSum(data, scale_idx)
}
\arguments{
\item{data}{a \code{\linkS4class{PROsetta_data}} object.}

\item{scale_idx}{the instrument index to obtain the raw sum scores.}
}
\description{
\code{\link{getScaleSum}} is a helper function for calculating instrument-wise raw sum scores from response data.
}
\examples{
getScaleSum(data_asq, 1)
getScaleSum(data_asq, 2)

}
