% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{getScaleSum}
\alias{getScaleSum}
\title{Calculate raw sum scores of a scale}
\usage{
getScaleSum(data, scale_idx)
}
\arguments{
\item{data}{a \code{\linkS4class{PROsetta_data}} object.}

\item{scale_idx}{the index of the scale to obtain the raw sum scores.}
}
\description{
\code{\link{getScaleSum}} is a helper function to calculate raw sum scores of a scale.
}
