% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\docType{class}
\name{PROsetta_data-class}
\alias{PROsetta_data-class}
\alias{loadData}
\title{An S4 class to represent \code{PROsetta} datasets.}
\usage{
loadData(config)
}
\arguments{
\item{config}{A PROsetta_config object. See \linkS4class{PROsetta_config}.}
}
\value{
An S4-class object containing the loaded data.
}
\description{
Loads data from the supplied configuration.
}
