% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\docType{methods}
\name{plotInfo}
\alias{plotInfo}
\alias{plotInfo,SingleGroupClass-method}
\title{Plot scale information}
\usage{
plotInfo(
  object,
  data,
  theta = seq(-4, 4, 0.1),
  t_score = FALSE,
  scale_label,
  color = "blue",
  lty = 1
)

\S4method{plotInfo}{SingleGroupClass}(
  object,
  data,
  theta = seq(-4, 4, 0.1),
  t_score = FALSE,
  scale_label,
  color = "blue",
  lty = 1
)
}
\arguments{
\item{object}{An \code{\linkS4class{SingleGroupClass}} object from \code{\link{runClassical}}.}

\item{data}{A \code{\linkS4class{PROsetta_data}} object.}

\item{theta}{Theta values to plot on the x-axis.}

\item{t_score}{Set to \code{TRUE} to convert thetas into T-scores.}

\item{scale_label}{Names of each scale.}

\item{color}{Line colors to plot.}

\item{lty}{Line types to plot.}
}
\description{
Plot scale information
}
