% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshiny.R
\name{PROsetta}
\alias{PROsetta}
\alias{guiPROsetta}
\title{PROsetta}
\usage{
PROsetta()

guiPROsetta()
}
\description{
\code{\link{PROsetta}} is a caller function to launch a Shiny app locally.
}
\examples{
if (interactive()) {
  PROsetta()
}

}
