% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\docType{class}
\name{PROsetta_config-class}
\alias{PROsetta_config-class}
\alias{createConfig}
\title{An S4 class to represent configurations for reading datasets and writing outputs.}
\usage{
createConfig(
  study_name = "Study",
  input_directory = getwd(),
  output_directory = getwd(),
  item_id = NULL,
  person_id = NULL,
  scale_id = NULL,
  response_file,
  itemmap_file,
  anchor_file,
  linking_method = "FIXEDPAR"
)
}
\arguments{
\item{study_name}{Character. A user-specified name of the study.}

\item{input_directory}{Character. The path of the input directory for reading dataset. Defaults to \code{getwd()}.}

\item{output_directory}{Character. The path of the output directory for writing outputs. Defaults to \code{getwd()}. A new folder is created at the path if it does not exist.}

\item{item_id}{Character. The name of the variable representing item IDs in \code{anchor_file} and \code{itemmap_file}. Guessed if left blank.}

\item{person_id}{Character. The name of the variable representing person IDs in \code{response_file}. Guessed if left blank.}

\item{scale_id}{Character. The name of the variable representing scale IDs in \code{itemmap_file}. Guessed if left blank.}

\item{response_file}{Character. The filename of a dataset containing IDs and the responses of the items.}

\item{itemmap_file}{Character. The filename for an item map.}

\item{anchor_file}{Character. The filename of a dataset containing the parameters of the items.}

\item{linking_method}{Character. The linking method to use. Accepts "\code{MM}", "\code{MS}", "\code{HB}", "\code{SL}", "\code{FIXEDPAR}", or "\code{NONE}".}
}
\value{
A \code{\linkS4class{PROsetta_config}} object.
}
\description{
Creates a config object for PROsetta
}
\section{Slots}{

\describe{
\item{\code{study_name}}{Character. A user-specified name of the study.}

\item{\code{input_directory}}{Character. The path of the input directory for reading dataset.}

\item{\code{output_directory}}{Character. The path of the output directory for writing outputs. A new folder is created at the path if it does not exist.}

\item{\code{item_id}}{Character. The name of the variable representing item IDs in \code{anchor_file} and \code{itemmap_file}.}

\item{\code{person_id}}{Character. The name of the variable representing person IDs in \code{response_file}.}

\item{\code{scale_id}}{Character. The name of the variable representing scale IDs in \code{itemmap_file}.}

\item{\code{response_file}}{Character. The filename of a dataset containing IDs and the responses of the items.}

\item{\code{itemmap_file}}{Character. The filename for an item map.}

\item{\code{anchor_file}}{Character. The filename of a dataset containing the parameters of the items.}

\item{\code{linking_method}}{Character. The linking method to use. Accepts "\code{MM}", "\code{MS}", "\code{HB}", "\code{SL}", "\code{FIXEDPAR}", or "\code{NONE}".}
}}

\examples{
\dontshow{
f1 <- tempfile()
f2 <- tempfile()
f3 <- tempfile()
write.csv(response_asq, f1, row.names = FALSE)
write.csv(itemmap_asq, f2, row.names = FALSE)
write.csv(anchor_asq, f3, row.names = FALSE)
cfg <- createConfig(response_file = f1, itemmap_file = f2, anchor_file = f3,
  linking_method = "FIXEDPAR")
}
\donttest{
write.csv(response_asq, "response.csv", row.names = FALSE)
write.csv(itemmap_asq, "itemmap.csv", row.names = FALSE)
write.csv(anchor_asq, "anchor.csv", row.names = FALSE)
cfg <- createConfig(
  response_file = "response.csv",
  itemmap_file = "itemmap.csv",
  anchor_file = "anchor.csv",
  linking_method = "FIXEDPAR"
)
}
\dontshow{
  file.remove(f1)
  file.remove(f2)
  file.remove(f3)
}
}
