% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{RunCalibration}
\alias{RunCalibration}
\title{Run Calibration}
\usage{
RunCalibration(Config, Data = NULL, ...)
}
\arguments{
\item{Config}{A PROsetta.Config object. See \code{\linkS4class{PROsetta.Config}}.}

\item{Data}{A PROsetta.Data object. See \code{\link{LoadData}} for loading a dataset.}

\item{...}{Additional arguments to pass onto \code{\link[mirt]{mirt}} in \href{https://CRAN.R-project.org/package=mirt}{\code{mirt}} package.}
}
\value{
An object containing item calibration results. This object can be used in \code{\link[mirt:coef-method]{coef}}, \code{\link[mirt]{itemfit}}, \code{\link[mirt]{itemplot}} in \href{https://CRAN.R-project.org/package=mirt}{\code{mirt}} package to extract wanted information.
}
\description{
Performs item calibration for the response data based on the supplied anchor information.
}
\examples{
\dontrun{
inputData = LoadData(new.Config)
outCalib = RunCalibration(new.Config, inputData)
mirt::coef(outCalib, IRTpars = TRUE, simplify = TRUE)
mirt::itemfit(outCalib, empirical.plot = 1)
mirt::itemplot(outCalib, item = 1, type = "info")
mirt::itemfit(outCalib, "S_X2", na.rm = TRUE)

new.Config@linkingMethod = "SL"
outCalib.Free = RunCalibration(new.Config, inputData)
outCalib.Free = RunCalibration(new.Config, inputData, technical = list(NCYCLES = 1000))
mirt::coef(outCalib.Free, IRTpars = TRUE, simplify = TRUE)
}
}
