% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{runClassical}
\alias{runClassical}
\title{Run a CTT-based reliability analysis}
\usage{
runClassical(config, data = NULL, omega = FALSE, ...)
}
\arguments{
\item{config}{A PROsetta_config object. See \code{\linkS4class{PROsetta_config}}.}

\item{data}{(Optional) A PROsetta_data object. See \code{\link{loadData}} for loading a dataset.}

\item{omega}{If TRUE, also obtains McDonald's omega using \code{\link[psych]{omega}} in \href{https://CRAN.R-project.org/package=psych}{\code{psych}} package.}

\item{...}{Additional arguments to pass onto \code{\link[psych]{omega}}.}
}
\value{
The results of reliability analysis.
}
\description{
Performs a Classial Test Theory (CTT) reliability analysis.
}
\examples{
classicalTable <- runClassical(cfg_asq)
classicalTable2 <- runClassical(cfg_asq, omega = TRUE) # also obtains omega
}
