% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{RunClassical}
\alias{RunClassical}
\title{Run a CTT-based reliability analysis}
\usage{
RunClassical(Config, Data = NULL, omega = FALSE, ...)
}
\arguments{
\item{Config}{A PROsetta.Config object. See \code{\linkS4class{PROsetta.Config}}.}

\item{Data}{A PROsetta.Data object. See \code{\link{LoadData}} for loading a dataset.}

\item{omega}{If TRUE, also obtains McDonald's omega using \code{\link[psych]{omega}} in \href{https://CRAN.R-project.org/package=psych}{\code{psych}} package.}

\item{...}{Additional arguments to pass onto \code{\link[psych]{omega}}.}
}
\value{
The results of reliability analysis.
}
\description{
Performs a Classial Test Theory (CTT) reliability analysis.
}
\examples{
\dontrun{
inputData = LoadData(new.Config)
classicalTable = RunClassical(new.Config, inputData)
classicalTable2 = RunClassical(new.Config, inputData, omega = TRUE)  # also obtains omega
}
}
