% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{runClassical}
\alias{runClassical}
\title{Run a CTT-based reliability analysis}
\usage{
runClassical(config, data = NULL, omega = FALSE, ...)
}
\arguments{
\item{config}{A PROsetta_config object. See \code{\linkS4class{PROsetta_config}}.}

\item{data}{(Optional) A PROsetta_data object. See \code{\link{loadData}} for loading a dataset.}

\item{omega}{If TRUE, also obtains McDonald's omega using \code{\link[psych]{omega}} in \href{https://CRAN.R-project.org/package=psych}{\code{psych}} package.}

\item{...}{Additional arguments to pass onto \code{\link[psych]{omega}}.}
}
\value{
The results of reliability analysis.
}
\description{
Performs a Classial Test Theory (CTT) reliability analysis.
}
\examples{
\dontshow{
f1 <- tempfile()
f2 <- tempfile()
f3 <- tempfile()
write.csv(response_asq, f1, row.names = FALSE)
write.csv(itemmap_asq, f2, row.names = FALSE)
write.csv(anchor_asq, f3, row.names = FALSE)
cfg <- createConfig(response_file = f1, itemmap_file = f2, anchor_file = f3)
}
\donttest{
write.csv(response_asq, "response.csv", row.names = FALSE)
write.csv(itemmap_asq, "itemmap.csv", row.names = FALSE)
write.csv(anchor_asq, "anchor.csv", row.names = FALSE)
cfg <- createConfig(
  response_file = "response.csv",
  itemmap_file = "itemmap.csv",
  anchor_file = "anchor.csv")
}
classicalTable <- runClassical(cfg)
classicalTable2 <- runClassical(cfg, omega = TRUE) # also obtains omega
\dontshow{
  file.remove(f1)
  file.remove(f2)
  file.remove(f3)
}
}
