% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{runDescriptive}
\alias{runDescriptive}
\title{Obtain a descriptive statistics table}
\usage{
runDescriptive(config, data = NULL)
}
\arguments{
\item{config}{A PROsetta_config object. See \code{\linkS4class{PROsetta_config}}.}

\item{data}{(Optional) A PROsetta_data object. See \code{\link{loadData}} for loading a dataset.}
}
\value{
A \code{data.frame} containing the descriptive statistics (mean, standard deviation, median, ...) of the variables in the dataset. These are calculated with \code{\link[psych]{describe}} in \href{https://CRAN.R-project.org/package=psych}{\code{psych}} package.
}
\description{
Obtains a table with descriptive statistics for each variable, from the supplied configuration and the dataset.
}
\examples{
descTable <- runDescriptive(cfg_asq)
}
