% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{RunDescriptive}
\alias{RunDescriptive}
\title{Obtain a descriptive statistics table}
\usage{
RunDescriptive(Config, Data = NULL)
}
\arguments{
\item{Config}{A PROsetta.Config object. See \code{\linkS4class{PROsetta.Config}}.}

\item{Data}{A PROsetta.Data object. See \code{\link{LoadData}} for loading a dataset.}
}
\value{
A \code{data.frame} containing the descriptive statistics (mean, standard deviation, median, ...) of the variables in the dataset. These are calculated with \code{\link[psych]{describe}} in \href{https://CRAN.R-project.org/package=psych}{\code{psych}} package.
}
\description{
Obtains a table with descriptive statistics for each variable, from the supplied configuration and the dataset.
}
\examples{
\dontrun{
inputData = LoadData(new.Config)
descTable = RunDescriptive(new.Config, inputData)
}
}
