% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preanalysis_functions.r
\name{runDescriptive}
\alias{runDescriptive}
\title{Obtain a descriptive statistics table}
\usage{
runDescriptive(data = NULL)
}
\arguments{
\item{data}{a \code{\linkS4class{PROsetta_data}} object. See \code{\link{loadData}} for loading a dataset.}
}
\value{
\code{\link{runDescriptive}} returns a \code{\link{data.frame}} containing descriptive statistics (mean, standard deviation, median, ...) of the items in the dataset. These are calculated with \code{\link[psych]{describe}} in '\href{https://CRAN.R-project.org/package=psych}{psych}' package.
}
\description{
\code{\link{runDescriptive}} is a descriptive function for obtaining descriptive statistics for each item in the dataset.
}
\examples{
out_desc <- runDescriptive(data_asq)

}
