% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{RunEquateObserved}
\alias{RunEquateObserved}
\title{Run Test Equating}
\usage{
RunEquateObserved(config, data = NULL, scaleTo = 1, scaleFrom = 2,
  type = "equipercentile", smooth = "loglinear", degrees = list(3,
  1), boot = TRUE, reps = 100, ...)
}
\arguments{
\item{config}{A PROsetta.Config object. See \code{\linkS4class{PROsetta.Config}}.}

\item{data}{A PROsetta.Data object. See \code{\link{LoadData}} for loading a dataset.}

\item{scaleTo}{Numeric. The index of the target scale to equate to. This and \code{scaleFrom} below both reference to the information stored in the \code{itemmap} slot of \code{Data} argument. The \code{scaleID} slot of \code{config} argument needs to be specified as the name of the varible containing the scale IDs in the \code{itemmap} slot.}

\item{scaleFrom}{Numeric. The index of the scale in need of test equating.}

\item{type}{The type of equating to be passed onto \code{\link[equate]{equate}} in \href{https://CRAN.R-project.org/package=equate}{\code{equate}} package.}

\item{smooth}{The type of smoothing method to be passed onto \code{\link[equate]{presmoothing}} in \href{https://CRAN.R-project.org/package=equate}{\code{equate}} package.}

\item{degrees}{The degrees of smoothing to be passed onto \code{\link[equate]{presmoothing}}.}

\item{boot}{Logical. Performs bootstrapping if \code{TRUE}.}

\item{reps}{Numeric. The number of replications in bootsrapping.}

\item{...}{Other arguments to pass onto \code{\link[equate]{equate}}..}
}
\value{
An \code{equate} object containing the test equating result. The printed summary statistics indicate the distributional properties of the two supplied scales and the equated scale. The rows titled \code{x} and \code{y} correspond to the scales specified in \code{scaleFrom} and \code{scaleTo} respectively. The row titled \code{yx} corresponds to the \code{scaleFrom} scale transformed to \code{scaleTo}. See \code{\link[equate]{equate}} for details.
}
\description{
Performs equipercentile test equating between two scales.
}
\examples{
\dontrun{
solution <- RunEquateObserved(cfg_asq, scaleTo = 1, scaleFrom = 2,
            type = "equipercentile", smooth = "loglinear")
}
}
