% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{runEquateObserved}
\alias{runEquateObserved}
\title{Run Test Equating}
\usage{
runEquateObserved(config, scale_to = 1, scale_from = 2,
  type = "equipercentile", smooth = "loglinear", degrees = list(3,
  1), boot = TRUE, reps = 100, data = NULL, ...)
}
\arguments{
\item{config}{A PROsetta_config object. See \code{\linkS4class{PROsetta_config}}.}

\item{scale_to}{Numeric. The index of the target scale to equate to. This and \code{scale_from} below both reference to the information stored in the \code{itemmap} slot of \code{Data} argument. The \code{scale_id} slot of \code{config} argument needs to be specified as the name of the varible containing the scale IDs in the \code{itemmap} slot.}

\item{scale_from}{Numeric. The index of the scale in need of test equating.}

\item{type}{The type of equating to be passed onto \code{\link[equate]{equate}} in \href{https://CRAN.R-project.org/package=equate}{\code{equate}} package.}

\item{smooth}{The type of smoothing method to be passed onto \code{\link[equate]{presmoothing}} in \href{https://CRAN.R-project.org/package=equate}{\code{equate}} package.}

\item{degrees}{The degrees of smoothing to be passed onto \code{\link[equate]{presmoothing}}.}

\item{boot}{Logical. Performs bootstrapping if \code{TRUE}.}

\item{reps}{Numeric. The number of replications in bootsrapping.}

\item{data}{A PROsetta_data object. See \code{\link{loadData}} for loading a dataset.}

\item{...}{Other arguments to pass onto \code{\link[equate]{equate}}..}
}
\value{
An \code{equate} object containing the test equating result. The printed summary statistics indicate the distributional properties of the two supplied scales and the equated scale. The rows titled \code{x} and \code{y} correspond to the scales specified in \code{scale_from} and \code{scale_to} respectively. The row titled \code{yx} corresponds to the \code{scale_from} scale transformed to \code{scale_to}. See \code{\link[equate]{equate}} for details.
}
\description{
Performs equipercentile test equating between two scales.
}
\examples{
\dontshow{
f1 <- tempfile()
f2 <- tempfile()
f3 <- tempfile()
write.csv(response_asq, f1, row.names = FALSE)
write.csv(itemmap_asq, f2, row.names = FALSE)
write.csv(anchor_asq, f3, row.names = FALSE)
cfg <- createConfig(response_file = f1, itemmap_file = f2, anchor_file = f3)
}
\donttest{
write.csv(response_asq, "response.csv", row.names = FALSE)
write.csv(itemmap_asq, "itemmap.csv", row.names = FALSE)
write.csv(anchor_asq, "anchor.csv", row.names = FALSE)
cfg <- createConfig(
  response_file = "response.csv",
  itemmap_file = "itemmap.csv",
  anchor_file = "anchor.csv")
}
solution <- runEquateObserved(cfg,
  scale_to = 1, scale_from = 2,
  type = "equipercentile", smooth = "loglinear"
)
\dontshow{
  file.remove(f1)
  file.remove(f2)
  file.remove(f3)
}
}
