% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{runFrequency}
\alias{runFrequency}
\title{Obtain a frequency table}
\usage{
runFrequency(data, check_frequency = TRUE)
}
\arguments{
\item{data}{A \code{\linkS4class{PROsetta_data}} object. See \code{\link{loadData}} for loading a dataset.}

\item{check_frequency}{Logical. If \code{TRUE}, check the frequency table for missing response categories, and display warning message if any is missing.}
}
\value{
A \code{data.frame} containing the frequency table of the dataset.
}
\description{
Obtains a frequency table from the supplied configuration and the dataset.
}
\examples{
out_freq <- runFrequency(data_asq)

}
