% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{runFrequency}
\alias{runFrequency}
\title{Obtain a frequency table}
\usage{
runFrequency(config, data = NULL, check_frequency = TRUE)
}
\arguments{
\item{config}{A PROsetta_config object. See \code{\linkS4class{PROsetta_config}}.}

\item{data}{A PROsetta_data object. See \code{\link{loadData}} for loading a dataset.}

\item{check_frequency}{Logical. If \code{TRUE}, check the frequency table for missing response categories, and display warning message if any is missing.}
}
\value{
A \code{data.frame} containing the frequency table of the dataset.
}
\description{
Obtains a frequency table from the supplied configuration and the dataset.
}
\examples{
freqTable <- runFrequency(cfg_asq)
}
