% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{RunFrequency}
\alias{RunFrequency}
\title{Obtain a frequency table}
\usage{
RunFrequency(config, data = NULL, checkFrequency = TRUE)
}
\arguments{
\item{config}{A PROsetta.Config object. See \code{\linkS4class{PROsetta.Config}}.}

\item{data}{A PROsetta.Data object. See \code{\link{LoadData}} for loading a dataset.}

\item{checkFrequency}{Logical. If \code{TRUE}, check the frequency table for missing response categories, and display warning message if any is missing.}
}
\value{
A \code{data.frame} containing the frequency table of the dataset.
}
\description{
Obtains a frequency table from the supplied configuration and the dataset.
}
\examples{
freqTable <- RunFrequency(cfg_asq)
}
