% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preanalysis_functions.r
\name{runFrequency}
\alias{runFrequency}
\title{Obtain a frequency table}
\usage{
runFrequency(data, check_frequency = TRUE)
}
\arguments{
\item{data}{a \code{\linkS4class{PROsetta_data}} object. See \code{\link{loadData}} for loading a dataset.}

\item{check_frequency}{if \code{TRUE}, check the frequency table for missing response categories, and display warning message if any is missing. (default = \code{TRUE})}
}
\value{
\code{\link{runFrequency}} returns a \code{\link{data.frame}} containing the frequency table.
}
\description{
\code{\link{runFrequency}} is a descriptive function for obtaining a frequency table from the dataset.
}
\examples{
freq_asq <- runFrequency(data_asq)
freq_dep <- runFrequency(data_dep)

}
