% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{RunLinking}
\alias{RunLinking}
\title{Run Scale Linking}
\usage{
RunLinking(config, data = NULL, ...)
}
\arguments{
\item{config}{A PROsetta.Config object. See \code{\linkS4class{PROsetta.Config}}.}

\item{data}{(Optional) A PROsetta.Data object. See \code{\link{LoadData}} for loading a dataset.}

\item{...}{Additional arguments to pass onto \code{\link[mirt]{mirt}} in \href{https://CRAN.R-project.org/package=mirt}{\code{mirt}} package.}
}
\value{
A list containing the scale linking results. These are obtained with \code{\link[plink]{plink-methods}} in \href{https://CRAN.R-project.org/package=plink}{\code{plink}} package.
}
\description{
Performs scale linking and obtains a set of transformation coefficients.
}
\examples{
\dontrun{
cfg_asq2 <- cfg_asq
cfg_asq2@linkingMethod = "SL"
solution <- RunLinking(cfg_asq2, technical = list(NCYCLES = 1000))
solution$link@constants$SL
}
}
