% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{runLinking}
\alias{runLinking}
\title{Run Scale Linking}
\usage{
runLinking(config, data = NULL, ...)
}
\arguments{
\item{config}{A PROsetta_config object. See \code{\linkS4class{PROsetta_config}}.}

\item{data}{(Optional) A PROsetta_data object. See \code{\link{loadData}} for loading a dataset.}

\item{...}{Additional arguments to pass onto \code{\link[mirt]{mirt}} in \href{https://CRAN.R-project.org/package=mirt}{\code{mirt}} package.}
}
\value{
A list containing the scale linking results. These are obtained with \code{\link[plink]{plink-methods}} in \href{https://CRAN.R-project.org/package=plink}{\code{plink}} package.
}
\description{
Performs scale linking and obtains a set of transformation coefficients.
}
\examples{
\dontrun{
cfg_asq2 <- cfg_asq
cfg_asq2@linking_method <- "SL"
solution <- runLinking(cfg_asq2, technical = list(NCYCLES = 1000))
solution$link@constants$SL
}

}
