% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{runRSSS}
\alias{runRSSS}
\title{Run Crosswalk Table Generation}
\usage{
runRSSS(
  data,
  ipar_linked,
  prior_mean = 0,
  prior_sd = 1,
  min_theta = -4,
  max_theta = 4,
  inc = 0.01,
  min_score = 1
)
}
\arguments{
\item{data}{A \code{\linkS4class{PROsetta_data}} object. See \code{\link{loadData}} for loading a dataset.}

\item{ipar_linked}{An object returned from \code{\link{runCalibration}} or \code{\link{runLinking}}.}

\item{prior_mean}{Prior mean.}

\item{prior_sd}{Prior standard deviation.}

\item{min_theta}{The lower limit of theta grid.}

\item{max_theta}{The upper limit of theta grid.}

\item{inc}{The increment to use in theta grid.}

\item{min_score}{Minimum item score (0 or 1).}
}
\value{
A list containing crosswalk tables.
}
\description{
Generates raw-score to standard-score crosswalk tables from supplied calibrated item parameters.
}
\examples{
out_link    <- runLinking(data_asq, method = "FIXEDPAR")
score_table <- runRSSS(data_asq, out_link)

}
