% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{RunRSSS}
\alias{RunRSSS}
\title{Run Scoring Table Generation}
\usage{
RunRSSS(config, data = NULL, calibration, priorMean = 0, priorSD = 1,
  minTheta = -4, maxTheta = 4, inc = 0.01, minScore = 1,
  Tscore = TRUE)
}
\arguments{
\item{config}{A PROsetta.Config object. See \code{\linkS4class{PROsetta.Config}}.}

\item{data}{(Optional) A PROsetta.Data object. See \code{\link{LoadData}} for loading a dataset.}

\item{calibration}{An object returned from \code{\link{RunCalibration}} or \code{\link{RunLinking}}}

\item{priorMean}{Prior mean.}

\item{priorSD}{Prior standard deviation.}

\item{minTheta}{LL of theta grid.}

\item{maxTheta}{UL of theta grid.}

\item{inc}{Increment of theta grid.}

\item{minScore}{Minimum item score (0 or 1).}

\item{Tscore}{TRUE to convert theta to Tscore.}
}
\value{
A list containing crosswalk tables.
}
\description{
Generates raw-score to scale-score crosswalk tables.
}
\examples{
solution <- RunCalibration(cfg_asq, inputData)
scoreTable = RunRSSS(outCalib)
outEquate = RunLinking(new.Config, inputData, technical = list(NCYCLES = 1000))
scoreTableLinking = RunRSSS(outEquate)
}
