% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPROsetta.R
\name{runRSSS}
\alias{runRSSS}
\title{Run Scoring Table Generation}
\usage{
runRSSS(
  config,
  data = NULL,
  calibration,
  priorMean = 0,
  priorSD = 1,
  minTheta = -4,
  maxTheta = 4,
  inc = 0.01,
  minScore = 1,
  Tscore = TRUE
)
}
\arguments{
\item{config}{A PROsetta_config object. See \code{\linkS4class{PROsetta_config}}.}

\item{data}{(Optional) A PROsetta_data object. See \code{\link{loadData}} for loading a dataset.}

\item{calibration}{An object returned from \code{\link{runCalibration}} or \code{\link{runLinking}}}

\item{priorMean}{Prior mean.}

\item{priorSD}{Prior standard deviation.}

\item{minTheta}{LL of theta grid.}

\item{maxTheta}{UL of theta grid.}

\item{inc}{Increment of theta grid.}

\item{minScore}{Minimum item score (0 or 1).}

\item{Tscore}{TRUE to convert theta to Tscore.}
}
\value{
A list containing crosswalk tables.
}
\description{
Generates raw-score to scale-score crosswalk tables.
}
\examples{
solution    <- runCalibration(cfg_asq)
score_table <- runRSSS(cfg_asq, calibration = solution)

cfg_asq2                <- cfg_asq
cfg_asq2@linking_method <- "SL"
solution                <- runLinking(cfg_asq2, technical = list(NCYCLES = 1000))
score_table_linking     <- runRSSS(cfg_asq2, calibration = solution)
}
