% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking_functions.R
\name{runRSSS}
\alias{runRSSS}
\title{Compute Crosswalk Tables}
\usage{
runRSSS(
  data,
  ipar_linked,
  prior_mean = 0,
  prior_sd = 1,
  min_theta = -4,
  max_theta = 4,
  inc = 0.05,
  min_score = 1
)
}
\arguments{
\item{data}{a \code{\linkS4class{PROsetta_data}} object. See \code{\link{loadData}} for loading a dataset.}

\item{ipar_linked}{an object returned from \code{\link{runLinking}} or \code{\link{runCalibration}}.}

\item{prior_mean}{prior mean. (default = \code{0.0})}

\item{prior_sd}{prior standard deviation. (default = \code{1.0})}

\item{min_theta}{the lower limit of theta quadratures for numerical integration. (default = \code{-4})}

\item{max_theta}{the upper limit of theta quadratures for numerical integration. (default = \code{4})}

\item{inc}{the increment between theta quadratures for numerical integration. (default = \code{0.05})}

\item{min_score}{minimum item score (0 or 1) for each scale (1, 2, and combined). If a single value is supplied, the value is applied to all scales. (default = \code{1})}
}
\value{
\code{\link{runRSSS}} returns a \code{\link{list}} containing crosswalk tables.
}
\description{
\code{\link{runRSSS}} is a function for generating raw-score to standard-score crosswalk tables from supplied calibrated item parameters.
}
\examples{
\donttest{
out_link    <- runLinking(data_asq, method = "FIXEDPAR")
score_table <- runRSSS(data_asq, out_link)
}

}
